+++ src/utils/index.ts
@@ -8,7 +8,7 @@ import graphqlProxy from './graphql_proxy';
 import safeCompare from './safe-compare';
 import storeSession from './store-session';
 import validateHmac from './hmac-validator';
-import validateShop, {sanitizeShop, sanitizeHost} from './shop-validator';
+import {sanitizeShop, sanitizeHost} from './shop-validator';
 import versionCompatible from './version-compatible';
 import withSession from './with-session';
 import getEmbeddedAppUrl from './get-embedded-app-url';
@@ -24,7 +24,6 @@ const ShopifyUtils = {
   safeCompare,
   storeSession,
   validateHmac,
-  validateShop,
   sanitizeShop,
   sanitizeHost,
   versionCompatible,