+++ src/utils/shop-validator.ts
@@ -2,19 +2,6 @@ import {InvalidHostError, InvalidShopError} from '../error';
 import {Context} from '../context';
 
 /**
- * Validates myshopify.com urls
- *
- * @param shop Shop url: {shop}.myshopify.com
- */
-export default function validateShop(shop: string): boolean {
-  console.warn(
-    'Deprecation notice: validateShop will be removed in the next major release.',
-  );
-  const shopUrlRegex = /^[a-zA-Z0-9][a-zA-Z0-9-]*\.myshopify\.(com|io)[/]*$/;
-  return shopUrlRegex.test(shop);
-}
-
-/**
  * Validates and sanitizes shop domain urls. If Context.CUSTOM_SHOP_DOMAINS is set, shops ending in those domains are
  * allowed. Accepts myshopify.com and myshopify.io by default.
  *